requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
face-recognition>=1.3.0
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
setuptools<81
